change file name

+++ ExplainabilityModularized/model_setup.py
@@ -21,7 +21,7 @@ def setup_model_tokenizer(is_quantized=False, model_name=None):
                                                     quantization_config=bnb_config, device_map=os.environ.get('CUDA_CORE'))
 
         else:
-            model = LlamaForCausalLM.from_pretrained(model_name, output_hidden_states=True,torch_dtype = torch.float16,
+            model = LlamaForCausalLM.from_pretrained(model_name, output_hidden_states=True,
                                                     device_map="auto")
 
         tokenizer = LlamaTokenizer.from_pretrained(model_name, add_special_tokens=False,
